backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-magic==0.4.27
opencv-python==4.8.1.78
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
httpx==0.25.2